requirements.txt
absl-py==0.10.0
astor==0.8.1
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
cycler==0.10.0
enum34==1.1.10
funcsigs==1.0.2
futures==3.3.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.4
MIDIUtil==1.2.1
mock==3.0.5
numpy==1.16.6
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-midi==0.2.4
pytz==2020.1
scipy==1.2.3
six==1.15.0
subprocess32==3.5.4
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
